Team name EDA
One of the things I took a notice of is the fact that team go through different names regularly as sponships change. So a team can be named "Sky Procycling" from 2011-2013, and then change to "Team Sky" from 2014-2019, and then "Team Ineos" from 2019-2020. All the same teams effectively with similar strategies, management, and collection of data and experience, but filed under different names in our classification model.

In this Notebook I investigate into the names of the 2021 teams, and change the previous names appropriately to the most current team name. This should improve the number of wins per team, and won't leave a newely named team as fresh first time team.